+++ lib/Dialect/TPU/Backend/include/backend/backend_tg_api.h
@@ -395,7 +395,7 @@ void cvi_backend_tg_concat_kernel(
     const CviBackendContext &ctx, uint32_t layer_id,
     int input_num, gaddr_t input_gaddrs[], gaddr_t output_gaddr,
     int axis_dims[], int concat_axis, int output_dim_size, int *output_dim,
-    bool do_relu, const int8_t *right_shift_width,
+    bool do_relu, const int *right_shift_width,
     const int *threshold_x_quantized, cvk_fmt_t fmt);
 
 void cvi_backend_tg_lut_kernel(const CviBackendContext &ctx, uint32_t layer_id,
